+++ RELEASE_NOTES.md
@@ -1,7 +1,6 @@
 ### 5.3.0-prerelease0002
-- experimental giraffe backend
-
-### 5.1.13
+- experimental giraffe backend### 5.1.14
+- [SimplePrimitives.fs] added textarea; js typo fix### 5.1.13
 - added orbit camera controller helpers
 
 ### 5.1.12